requirements.txt
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
mtcnn==0.1.1
scikit-learn==1.3.2
matplotlib==3.7.3
Pillow==10.0.1
streamlit==1.28.2
plotly==5.17.0
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
scipy==1.11.4
tqdm==4.66.1
